LayoutTests/fast/events/resize-event-not-fired-during-page-scale-expected.txt

Tests that changing page scale factor does not fire a spurious resize event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resizeCount is 0
PASS successfullyParsed is true

TEST COMPLETE
